logger: move tint package into logger package
