+++ internal/logger/tint/buffer.go
