+++ internal/logger/tint/doc.go
